Refactor to run custom notifications in worker process
